Update RLW_arrange_channels.m
In this way, channels are selected based on the label and not on the index in the data matrix. This solves the problem that appears when a group of datasets have different number of channels (e.g. when a noisy channel was removed in some but not all datasets). In the previous version, the script would look in the indexes of the first dataset, which could match or not the indexes in the rest of the selected datasets.